Search and trips

+++ src/main/java/com/narwal/trainservice/service/TripsService.java
@@ -1,5 +1,6 @@
 package com.narwal.trainservice.service;
 
+import com.narwal.trainservice.model.Train;
 import com.narwal.trainservice.model.Trip;
 import com.narwal.trainservice.repository.TripsRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,20 +15,26 @@ public class TripsService {
     @Autowired
     TripsRepo tripsRepo;
 
-    public Trip createTrip(Trip trip){
-        return tripsRepo.save(trip);
+    public Optional<Trip> createTrip(Trip trip){
+        return Optional.of(tripsRepo.save(trip));
     }
 
-    public Trip updateTrip(String tripId, Trip trip){
+    public Optional<Trip> updateTrip(String tripId, Trip trip){
         Optional<Trip> tripData = tripsRepo.findByTripId(tripId);
         if(tripData.isPresent()){
-            return tripsRepo.save(trip);
+            trip.setTripId(tripId);
+            return Optional.of(tripsRepo.save(trip));
         }
-        return trip;
+        return Optional.empty();
     }
 
-    public void deleteTrip(String tripId){
-        tripsRepo.deleteByTripId(tripId);
+    public Optional<Trip> deleteTrip(String tripId){
+        Optional<Trip> trip = tripsRepo.findById(tripId);
+        if(trip.isPresent()){
+            tripsRepo.deleteByTripId(tripId);
+            return trip;
+        }
+        return Optional.empty();
     }
 
     public Optional<Trip> getTrip(String tripId){
@@ -36,5 +43,9 @@ public class TripsService {
 
     public List<Trip> getAllTrips(){
         return tripsRepo.findAll();
+    }
+
+    public List<Trip> getAlTripsBySrcAndDest(String srcStationCode, String destStationCode) {
+        return tripsRepo.findBySourceStationCodeAndDestinationStationCode(srcStationCode, destStationCode);
     }
 }
